+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - fixed security issue with the newly introduced experimental "UseCreateToken=y" machanism
+- fixed issue with "UseCreateToken=y" when using a MSFT online account
+